refactor: TestMemberBuilder를 support 패키지로 옮기고, 오픈된 생성자를 사용해 객체를 만들도록 변경
