+++ backend/src/test/java/edonymyeon/backend/consumption/repository/ConsumptionRepositoryTest.java
@@ -2,7 +2,7 @@ package edonymyeon.backend.consumption.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import edonymyeon.backend.member.domain.TestMemberBuilder;
+import edonymyeon.backend.support.TestMemberBuilder;
 import edonymyeon.backend.support.IntegrationTest;
 import edonymyeon.backend.consumption.domain.Consumption;
 import edonymyeon.backend.consumption.domain.ConsumptionType;
