+++ backend/src/test/java/edonymyeon/backend/setting/domain/SettingTest.java
@@ -3,7 +3,7 @@ package edonymyeon.backend.setting.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import edonymyeon.backend.member.domain.Member;
-import edonymyeon.backend.member.domain.TestMemberBuilder;
+import edonymyeon.backend.support.TestMemberBuilder;
 import java.util.Collections;
 import java.util.List;
 import org.junit.jupiter.api.Test;
@@ -16,9 +16,6 @@ class SettingTest {
     void 설정_항목을_체크하여_OnOff할_수_있다() {
         final Member member = testMemberBuilder.builder()
                 .id(1L)
-                .email("test@gmail.com")
-                .password("password123")
-                .nickname("nickName")
                 .buildWithoutSaving();
         final Setting setting
                 = new Setting(member, SettingType.NOTIFICATION_PER_THUMB, Collections.emptyList(),
@@ -33,9 +30,6 @@ class SettingTest {
     void 상위_항목을_disable하는_경우_하위_항목도_함께_disable된다() {
         final Member member = testMemberBuilder.builder()
                 .id(1L)
-                .email("test@gmail.com")
-                .password("password123")
-                .nickname("nickName")
                 .buildWithoutSaving();
 
         final Setting 하위항목1 = new Setting(member, SettingType.NOTIFICATION_PER_THUMB, Collections.emptyList(),
@@ -57,9 +51,6 @@ class SettingTest {
     void 상위_항목에_의해_disable된_항목이_상위_항목에_의해_다시_activated된_경우_이전의_설정을_다시_복구한다() {
         final Member member = testMemberBuilder.builder()
                 .id(1L)
-                .email("test@gmail.com")
-                .password("password123")
-                .nickname("nickName")
                 .buildWithoutSaving();
 
         final Setting 하위항목1 = new Setting(member, SettingType.NOTIFICATION_PER_THUMB, Collections.emptyList(),
