+++ backend/src/test/java/edonymyeon/backend/thumbs/domain/AllThumbsInPostTest.java
@@ -5,7 +5,7 @@ import static edonymyeon.backend.thumbs.domain.ThumbsType.UP;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 import edonymyeon.backend.member.domain.Member;
-import edonymyeon.backend.member.domain.TestMemberBuilder;
+import edonymyeon.backend.support.TestMemberBuilder;
 import edonymyeon.backend.post.domain.Post;
 import java.util.Collections;
 import java.util.List;
